# src/message/schemas/message.schema.ts
import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
import { Document, Types } from 'mongoose';

@Schema({ timestamps: true })
export class Message extends Document {

  @Prop({ type: Types.ObjectId, ref: 'Utilisateur', required: true })
  senderId: Types.ObjectId;

  @Prop({ type: Types.ObjectId, ref: 'Utilisateur', required: true })
  receiverId: Types.ObjectId;

  @Prop({ required: true })
  content: string;

  @Prop({ default: 'text' })
  type: string;

  @Prop({ default: false })
  isRead: boolean;

  @Prop({ type: Date, default: null })
  readAt: Date | null;

  // 👇👇 AJOUT OBLIGATOIRE pour enlever l’erreur
  @Prop({ type: Date })
  createdAt: Date;

  @Prop({ type: Date })
  updatedAt: Date;
  @Prop({
  type: [
    {
      userId: { type: Types.ObjectId, ref: 'Utilisateur' },
      emoji: { type: String }
    }
  ],
  default: []
})
reactions: {
  userId: Types.ObjectId;
  emoji: string;
}[];

}

export const MessageSchema = SchemaFactory.createForClass(Message);